requirements.txt
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.7
Django==5.0
django-extensions==3.2.3
djangorestframework==3.14.0
filelock==3.13.1
flake8==6.1.0
identify==2.5.32
idna==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pre-commit==3.5.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
virtualenv==20.25.0
Pillow==10.1.0
